+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the "codeboost" extension will be documented in this file.
 
+## [1.1] 2025-04-26
+- 登録した Main Marker, Test Marker, Line Marker, Comment Marker をコンテクストメニューで呼び出す機能を追加
+- Added functionality to invoke the registered Main Marker, Test Marker, Line Marker, and Comment Marker from the context menu.
+
 ## [1.0.2] 2024-09-01
 - 検索フォルダ内のサブフォルダが引っかからないよう修正